fix: bigint

+++ db/migrate/20180515112421_create_review_comments.rb
@@ -7,8 +7,8 @@ class CreateReviewComments < ActiveRecord::Migration[5.1]
       t.text :body
       t.string :path
       t.integer :position
-      t.integer :github_id
-      t.integer :in_reply_to_id
+      t.bigint :github_id
+      t.bigint :in_reply_to_id
       t.integer :status
       t.timestamp :github_created_at
       t.timestamp :github_updated_at
